+++ src/client/microsoftAuth.js
@@ -11,6 +11,7 @@ function validateOptions (options) {
   if (options.authTitle === undefined) {
     options.authTitle = Titles.MinecraftNintendoSwitch
     options.deviceType = 'Nintendo'
+    options.flow = 'live'
   }
 }