+++ src/index.d.ts
@@ -53,7 +53,7 @@ declare module 'minecraft-protocol' {
 		once(event: 'sessionce', handler: (sessionce: any) => PromiseLike): this
 		once(event: 'state', handler: (newState: States, oldState: States) => PromiseLike): this
 		once(event: 'end', handler: (reasonce: string) => PromiseLike): this
-		once(event: 'concenect', handler: () => PromiseLike): this
+		once(event: 'connect', handler: () => PromiseLike): this
 		once(event: string, handler: (data: any, packetMeta: PacketMeta) => PromiseLike): this
 		once(event: `raw.${string}`, handler: (buffer: Buffer, packetMeta: PacketMeta) => PromiseLike): this
 	}
@@ -132,7 +132,7 @@ declare module 'minecraft-protocol' {
 		validateChannelProtocol?: boolean,
 		realms?: RealmsOptions
 		// 1.19+
-		disableChatSigning: boolean
+		disableChatSigning?: boolean
 	}
 
 	export class Server extends EventEmitter {
@@ -182,7 +182,7 @@ declare module 'minecraft-protocol' {
 		validateChannelProtocol?: boolean
 		// 1.19+
 		// Require connecting clients to have chat signing support enabled
-		enforceSecureProfile: boolean
+		enforceSecureProfile?: boolean
 	}
 
 	export interface SerializerOptions {